[#84] Move dagreIcons to constants folder
