+++ src/pages/ScriptPage.jsx
@@ -12,8 +12,8 @@ import SFormsModal from "../../components/sform/SFormsModal.jsx";
 import ModuleTypesSelection from "../../components/ModuleTypesSelection.jsx";
 import ScriptInputOutputModal from "../../components/sform/ScriptInputOutputModal.jsx";
 import { Dropdown } from "semantic-ui-react";
-import { ICONS_MAP } from "./DagreIcons.jsx";
-import ScriptFunctionSelection from "../../components/ScriptFunctionSelection.jsx";
+import { ICONS_MAP } from "../constants/dagreIcons.js";
+import ScriptFunctionSelection from "../components/ScriptFunctionSelection.jsx";
 import { Button } from "react-bootstrap";
 import ValidationReportModal from "../../components/modal/ValidationReportModal.jsx";
 import MoveModuleModal from "../../components/modal/MoveModuleModal.jsx";
